add plugin docs
