+++ docs/autodoc_build/eido.md
@@ -123,14 +123,77 @@ Get a list of available target formats
 
 
 ```python
-def convert_project(prj, target_format, plugin_kwargs={})
+def convert_project(prj, target_format, plugin_kwargs=None)
 ```
 
 Convert a `peppy.Project` object to a selected format
 #### Parameters:
 
 - `prj` (`peppy.Project`):  a Project object to convert
+- `plugin_kwargs` (`dict`):  kwargs to pass to the plugin function
 - `target_format` (`str`):  the format to convert the Project object to
+
+
+#### Raises:
+
+- `EidoFilterError`:  if the requested filter is not defined
+
+
+
+
+```python
+def basic_pep_filter(p, **kwargs)
+```
+
+Basic PEP filter, that does not convert the Project object.
+
+This filter can save the PEP representation to file, if kwargs include `path`.
+#### Parameters:
+
+- `p` (`peppy.Project`):  a Project to run filter on
+
+
+
+
+```python
+def yaml_pep_filter(p, **kwargs)
+```
+
+YAML PEP filter, that returns Project object representation.
+
+This filter can save the YAML to file, if kwargs include `path`.
+#### Parameters:
+
+- `p` (`peppy.Project`):  a Project to run filter on
+
+
+
+
+```python
+def csv_pep_filter(p, **kwargs)
+```
+
+CSV PEP filter, that returns Sample object representations
+
+This filter can save the CSVs to files, if kwargs include
+`sample_table_path` and/or `subsample_table_path`.
+#### Parameters:
+
+- `p` (`peppy.Project`):  a Project to run filter on
+
+
+
+
+```python
+def yaml_samples_pep_filter(p, **kwargs)
+```
+
+YAML samples PEP filter, that returns only Sample object representations.
+
+This filter can save the YAML to file, if kwargs include `path`.
+#### Parameters:
+
+- `p` (`peppy.Project`):  a Project to run filter on
