+++ docs/plugin_api_docs.md
@@ -1,0 +1,94 @@
+Final targets: basic_pep_filter, csv_pep_filter, yaml_pep_filter, yaml_samples_pep_filter
+<script>
+document.addEventListener('DOMContentLoaded', (event) => {
+  document.querySelectorAll('h3 code').forEach((block) => {
+    hljs.highlightBlock(block);
+  });
+});
+</script>
+
+<style>
+h3 .content {
+    padding-left: 22px;
+    text-indent: -15px;
+ }
+h3 .hljs .content {
+    padding-left: 20px;
+    margin-left: 0px;
+    text-indent: -15px;
+    martin-bottom: 0px;
+}
+h4 .content, table .content, p .content, li .content { margin-left: 30px; }
+h4 .content {
+    font-style: italic;
+    font-size: 1em;
+    margin-bottom: 0px;
+}
+
+</style>
+
+
+# Built-in `eido` Plugins Documentation
+
+
+```python
+def basic_pep_filter(p, **kwargs)
+```
+
+Basic PEP filter, that does not convert the Project object.
+
+This filter can save the PEP representation to file, if kwargs include `path`.
+#### Parameters:
+
+- `p` (`peppy.Project`):  a Project to run filter on
+
+
+
+
+```python
+def yaml_pep_filter(p, **kwargs)
+```
+
+YAML PEP filter, that returns Project object representation.
+
+This filter can save the YAML to file, if kwargs include `path`.
+#### Parameters:
+
+- `p` (`peppy.Project`):  a Project to run filter on
+
+
+
+
+```python
+def csv_pep_filter(p, **kwargs)
+```
+
+CSV PEP filter, that returns Sample object representations
+
+This filter can save the CSVs to files, if kwargs include
+`sample_table_path` and/or `subsample_table_path`.
+#### Parameters:
+
+- `p` (`peppy.Project`):  a Project to run filter on
+
+
+
+
+```python
+def yaml_samples_pep_filter(p, **kwargs)
+```
+
+YAML samples PEP filter, that returns only Sample object representations.
+
+This filter can save the YAML to file, if kwargs include `path`.
+#### Parameters:
+
+- `p` (`peppy.Project`):  a Project to run filter on
+
+
+
+
+
+
+
+*Version Information: `eido` v0.1.5-dev, generated by `lucidoc` v0.4.2*
