+++ eido/__init__.py
@@ -5,6 +5,7 @@ Project configuration
 from ._version import __version__
 from .const import *
 from .conversion import *
+from .conversion_plugins import *
 from .exceptions import *
 from .inspection import *
 from .schema import *
@@ -18,4 +19,8 @@ __all__ = [
     "inspect_project",
     "get_available_pep_filters",
     "convert_project",
+    "basic_pep_filter",
+    "yaml_pep_filter",
+    "csv_pep_filter",
+    "yaml_samples_pep_filter",
 ]
