+++ mkdocs.yml
@@ -5,7 +5,7 @@ repo_url: http://github.com/pepkit/eido
 pypi_name: eido
 
 nav:
-  - Getting Started: 
+  - Getting Started:
     - Introduction: README.md
     - Install and configure: install.md
   - How-to Guides:
@@ -17,6 +17,7 @@ nav:
   - Reference:
     - Example schemas: example-schemas.md
     - Eido Python API: autodoc_build/eido.md
+    - Built-in filters API: plugin_api_docs.md
     - Support: https://github.com/pepkit/eido/issues
     - Contributing: contributing.md
     - Changelog: changelog.md
@@ -27,5 +28,6 @@ plugins:
   - databio:
       autodoc_build: "docs/autodoc_build"
       autodoc_package: "eido"
+      blacklist: "basic_pep_filter,yaml_pep_filter,csv_pep_filter,yaml_samples_pep_filter"
       no_top_level: true
   - search
