+++ update_plugins_api_docs.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+lucidoc eido -P rst --whitelist basic_pep_filter,yaml_pep_filter,csv_pep_filter,yaml_samples_pep_filter > docs/plugin_api_docs.md
